pyproject.toml
[project]
name = "inxtinct"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.13.3",
    "httpx>=0.28.1",
    "langchain>=1.1.2",
    "langchain-core>=1.1.1",
    "langchain-groq>=1.1.0",
    "langchain-mcp-adapters>=0.1.14",
    "pydantic>=2.12.5",
    "pydantic-core>=2.41.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "regex>=2025.11.3",
    "streamlit>=1.52.1",
]
